finished

+++ courseservice/src/main/java/com/csye6225/spring2020/courseservice/service/ProfessorsService.java
@@ -1,9 +1,7 @@
 package com.csye6225.spring2020.courseservice.service;
 
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.List;
+import java.text.SimpleDateFormat;
+import java.util.*;
 
 //import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBMapper;
 //import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBQueryExpression;
@@ -12,20 +10,22 @@ import java.util.List;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
 import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBMapper;
 import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBQueryExpression;
+import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBSaveExpression;
 import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBScanExpression;
 import com.amazonaws.services.dynamodbv2.model.AttributeValue;
+import com.amazonaws.services.dynamodbv2.model.ConditionalCheckFailedException;
+import com.amazonaws.services.dynamodbv2.model.ExpectedAttributeValue;
+import com.csye6225.spring2020.courseservice.datamodel.Professor;
 import com.csye6225.spring2020.courseservice.datamodel.DynamoDBConnector;
 import com.csye6225.spring2020.courseservice.datamodel.InMemoryDatabase;
 import com.csye6225.spring2020.courseservice.datamodel.Professor;
 
 public class ProfessorsService {
-
-    private static HashMap<String, Professor> prof_Map = InMemoryDatabase.getProfessorDB();
     private DynamoDBMapper mapper;
     private static AmazonDynamoDB client;
 
     public ProfessorsService() {
-        client = DynamoDBConnector.getClient(false);
+        client = DynamoDBConnector.getClient(true);
         mapper = new DynamoDBMapper(client);
     }
 
@@ -34,51 +34,84 @@ public class ProfessorsService {
         return list;
     }
 
-    public Professor addProfessor(Professor prof) {
-        mapper.save(prof);
-        return prof;
-    }
-
     // Getting One Professor
     public Professor getProfessor(String profId) {
         Professor prof = mapper.load(Professor.class, profId);
+        if(prof==null){
+            System.out.println("This professor is not exist: Invalid id:" + profId);
+        }
+        return prof;
+    }
+    public Professor addProfessor(Professor prof) {
+        //check if prof is valid
+        //1. professorId should be unique
+        try {
+            DynamoDBSaveExpression saveExpression = new DynamoDBSaveExpression();
+            HashMap<String, ExpectedAttributeValue> expected = new HashMap();
+            expected.put("professorId", new ExpectedAttributeValue().withExists(false));
+            saveExpression.setExpected(expected);
+            //Set post date
+            prof.setJoiningDate(getDate());
+            mapper.save(prof, saveExpression);
+        } catch (ConditionalCheckFailedException e) {
+            //if professor table doesn't contain this professor, throw exception
+            System.out.println("This professor is exist: Invalid professor id:" + prof.getProfessorId());
+            return null;
+        }
         return prof;
     }
 
-    // Deleting a professor
-    public Professor deleteProfessor(String profId) {
-        Professor oldObject = getProfessor(profId);
-        mapper.delete(oldObject);
+    // Deleting a Professor
+    public Professor deleteProfessor(String Id) {
+        Professor oldObject = getProfessor(Id);
+        if (oldObject != null) {
+            mapper.delete(oldObject);
+        } else {
+            System.out.println("Delete Fail");
+        }
         return oldObject;
     }
 
     // Updating Professor Info
-    public Professor updateProfessorInformation(String profId, Professor prof) {
-//		if (oldProfObject == null)
-//        {
-//            return null;
-//        }
-        prof.setProfessorId(profId);
-        mapper.save(prof);
+    public Professor updateProfessor(String Id, Professor prof) {
+        prof.setProfessorId(Id);
+        //check if professor table contains this professorId before update
+        try {
+            DynamoDBSaveExpression saveExpression = new DynamoDBSaveExpression();
+            HashMap<String, ExpectedAttributeValue> expected = new HashMap();
+            expected.put("professorId", new ExpectedAttributeValue(new AttributeValue(Id)));
+            saveExpression.setExpected(expected);
+            mapper.save(prof, saveExpression);
+        } catch (ConditionalCheckFailedException e) {
+            //if professor table doesn't contain this professor,
+            System.out.println("This professor is not exist: Invalid id:" + Id);
+            return null;
+        }
         return prof;
     }
-
     // Get professors in a department
     public List<Professor> getProfessorsByDepartment(String department) {
         HashMap<String, AttributeValue> eav = new HashMap<>();
         eav.put(":value1", new AttributeValue().withS(department));
         DynamoDBQueryExpression<Professor> queryExpression = new DynamoDBQueryExpression()
-                .withIndexName("department")
+                .withIndexName("department-index")
                 .withKeyConditionExpression("department = :value1")
-                .withExpressionAttributeValues(eav);
-        List<Professor> list = mapper.query(Professor.class, queryExpression);
+                .withExpressionAttributeValues(eav)
+                .withConsistentRead(false);
+        List<Professor> list=new LinkedList<>();
+        try{
+             list= mapper.query(Professor.class, queryExpression);
+        }catch (ConditionalCheckFailedException e) {
+            System.out.println(e);
+        }
         return list;
     }
-
-    public boolean isExist(String profId) {
-        return prof_Map.containsKey(profId);
+    private String getDate() {
+        String pattern = "yyyy-MM-dd";
+        SimpleDateFormat simpleDateFormat = new SimpleDateFormat(pattern);
+        String date = simpleDateFormat.format(new Date());
+        return date;
     }
-
 
     // Get professors for a year with a size limit
